+++ common/src/main/java/org/wso2/testgrid/common/TestScenario.java
@@ -75,7 +75,7 @@ public class TestScenario extends AbstractUUIDEntity implements Serializable {
     @Column(name = "is_pre_script_success")
     private boolean isPreScriptSuccessful = false;
 
-    @Column(name = "configChangeSet_name")
+    @Column(name = "config_change_set_name")
     private String configChangeSetName;
 
     @Column(name = "is_post_script_success")
