+++ core/src/main/java/org/wso2/testgrid/core/TestPlanExecutor.java
@@ -20,14 +20,9 @@ package org.wso2.testgrid.core;
 
 import com.google.gson.Gson;
 import org.apache.commons.lang.StringUtils;
-import org.apache.hc.client5.http.fluent.Content;
 import org.apache.hc.client5.http.fluent.Request;
 import org.apache.hc.client5.http.fluent.Response;
-import org.apache.hc.core5.http.ContentType;
 import org.apache.http.HttpHeaders;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.testgrid.automation.exception.ReportGeneratorException;
@@ -60,6 +55,8 @@ import org.wso2.testgrid.common.exception.UnsupportedDeployerException;
 import org.wso2.testgrid.common.exception.UnsupportedProviderException;
 import org.wso2.testgrid.common.util.DataBucketsHelper;
 import org.wso2.testgrid.common.util.StringUtil;
+import org.wso2.testgrid.core.configchangeset.ConfigChangeSetExecutor;
+import org.wso2.testgrid.core.configchangeset.ConfigChangeSetFactory;
 import org.wso2.testgrid.core.exception.ScenarioExecutorException;
 import org.wso2.testgrid.core.exception.TestPlanExecutorException;
 import org.wso2.testgrid.dao.TestGridDAOException;
@@ -73,10 +70,7 @@ import org.wso2.testgrid.tinkerer.exception.TinkererOperationException;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -319,19 +313,28 @@ public class TestPlanExecutor {
             List<ConfigChangeSet> configChangeSetList = testPlan.getScenarioConfig().getConfigChangeSets();
             // Run test with config change sets
             if (configChangeSetList != null) {
-                if (initConfigChangeSet(testPlan)) {
-                    for (ConfigChangeSet configChangeSet : configChangeSetList) {
-                        for (TestScenario testScenario : testPlan.getTestScenarios()) {
-                            if (configChangeSet.getName().equals(testScenario.getConfigChangeSetName())) {
-                                if (applyConfigChangeSet(testPlan, configChangeSet, true)) {
-                                    executeTestScenario(testScenario, deploymentCreationResult, testPlan);
-                                    applyConfigChangeSet(testPlan, configChangeSet, false);
+                OSCategory osCategory = getOSCatagory(testPlan.getInfraParameters());
+                Optional<ConfigChangeSetExecutor> configChangeSetExecutor =
+                        ConfigChangeSetFactory.getExecuter(osCategory);
+                if (configChangeSetExecutor.isPresent()) {
+
+                    if (configChangeSetExecutor.get().initConfigChangeSet(testPlan)) {
+                        for (ConfigChangeSet configChangeSet : configChangeSetList) {
+                            for (TestScenario testScenario : testPlan.getTestScenarios()) {
+                                if (configChangeSet.getName().equals(testScenario.getConfigChangeSetName())) {
+                                    if (configChangeSetExecutor.get().
+                                            applyConfigChangeSet(testPlan, configChangeSet, true)) {
+                                        executeTestScenario(testScenario, deploymentCreationResult, testPlan);
+                                        configChangeSetExecutor.get().
+                                                applyConfigChangeSet(testPlan, configChangeSet, false);
+                                    }
                                 }
                             }
                         }
-
+                        configChangeSetExecutor.get().deInitConfigChangeSet(testPlan);
                     }
-                    deInitConfigChangeSet(testPlan);
+                } else {
+                    logger.error("Unable to find a Tinker Executor for OS category " + osCategory);
                 }
             } else {
                 // Run test without config change set
@@ -384,139 +387,6 @@ public class TestPlanExecutor {
             logger.error(StringUtil.concatStrings(
                     "Error occurred while persisting test scenario ", testScenario.getName()), e);
         }
-    }
-
-    /**
-     * Apply config change set script before and after run test scenarios
-     *
-     * @param testPlan the test plan
-     * @param configChangeSet   config change set
-     * @param isInit run apply config-script if true. else, run revert-config script
-     * @return execution passed or failed
-     */
-    private boolean applyConfigChangeSet(TestPlan testPlan, ConfigChangeSet configChangeSet, boolean isInit) {
-        try {
-            URL configChangeSetRepoPath = new URL(testPlan.getConfigChangeSetRepository());
-            String filePath = configChangeSetRepoPath.getPath();
-            String fileName = filePath.substring(filePath.lastIndexOf('/') + 1);
-            String shellCommand = "./repos/" + fileName + "-master/config-sets/" + configChangeSet.getName();
-            if (isInit) {
-                shellCommand = shellCommand.concat("/apply-config.sh &>/dev/null");
-            } else {
-                shellCommand = shellCommand.concat("/revert-config.sh &>/dev/null");
-            }
-            // Generate array of commands that need to be applied
-            String[] applyShellCommand = { shellCommand };
-            return applyShellCommandOnAgent(testPlan, applyShellCommand);
-        } catch (MalformedURLException e) {
-            logger.warn("Error parsing scenario path " + e);
-            return false;
-        }
-    }
-
-    /**
-     * Initialize agent before running config change set
-     *
-     * @param testPlan  The test plan
-     * @return          True if execution success. Else, false
-     */
-    private boolean initConfigChangeSet(TestPlan testPlan) {
-        try {
-            URL configChangeSetRepoPath = new URL(testPlan.getConfigChangeSetRepository());
-            String filePath = configChangeSetRepoPath.getPath();
-            String fileName = filePath.substring(filePath.lastIndexOf('/') + 1);
-            String[] initShellCommand = {
-                    "mkdir repos",
-                    "cd repos && curl -LJO " +
-                            testPlan.getConfigChangeSetRepository() + "/archive/master.tar.gz   &>/dev/null",
-                    "cd repos && tar xvzf " + fileName + "-master.tar.gz  &>/dev/null",
-                    "chmod -R 755 repos/" + fileName + "-master/config-sets/ &>/dev/null"
-            };
-            return applyShellCommandOnAgent(testPlan, initShellCommand);
-        } catch (MalformedURLException e) {
-            logger.warn("Error parsing scenario path " + e);
-            return false;
-        }
-    }
-
-    /**
-     * Revert back changes did in initConfigChangeSet
-     *
-     * @param testPlan  The test plan
-     * @return          True if execution success. Else, false
-     */
-    private boolean deInitConfigChangeSet(TestPlan testPlan) {
-        String[] deInitShellCommand = {
-                "rm -rf repos"
-        };
-        return applyShellCommandOnAgent(testPlan, deInitShellCommand);
-    }
-
-    /**
-     * Send an array of shell command to an agent
-     *
-     * @param testPlan          The test plan
-     * @param initShellCommand  Array of shell command
-     * @return  True if execution success. Else, false
-     */
-    private boolean applyShellCommandOnAgent(TestPlan testPlan, String[] initShellCommand) {
-        String tinkererHost = ConfigurationContext.getProperty(
-                ConfigurationContext.ConfigurationProperties.DEPLOYMENT_TINKERER_REST_BASE_PATH);
-        // Get authentication detail from config.properties and encode with BASE64
-        String authenticationString = ConfigurationContext.getProperty(
-                ConfigurationContext.ConfigurationProperties.DEPLOYMENT_TINKERER_USERNAME) + ":" +
-                ConfigurationContext.getProperty(
-                        ConfigurationContext.ConfigurationProperties.DEPLOYMENT_TINKERER_PASSWORD);
-        String authenticationToken = "Basic " + Base64.getEncoder().encodeToString(
-                authenticationString.getBytes(StandardCharsets.UTF_8));
-        try {
-            Content agentResponse = Request.Get(tinkererHost + "/agents")
-                    .addHeader(HttpHeaders.AUTHORIZATION, authenticationToken).
-                            execute().returnContent();
-            try {
-                // Todo use with Agent class with jackson mapper
-                JSONArray agentArray = new JSONArray(agentResponse.asString());
-                for (int index = 0; index < agentArray.length(); index++) {
-                    JSONObject agent = agentArray.getJSONObject(index);
-                    if (agent.getString("testPlanId").equals(testPlan.getId())) {
-                        for (String shellCommand : initShellCommand) {
-                            logger.info("exec " + shellCommand);
-                            Content shellCommandResponse = sendShellCommand(tinkererHost, authenticationToken
-                                    , agent, shellCommand);
-                        }
-                    }
-                }
-                return true;
-            } catch (JSONException e) {
-                logger.warn("JSON parsing error " + e);
-                return false;
-            }
-        } catch (IOException e) {
-            logger.warn("Error in API call " + e);
-            return false;
-        }
-    }
-
-    /**
-     * Send a shell command to the agent and execute it
-     *
-     * @param tinkererHost          Tinkerer host address
-     * @param authenticationKey     Tinkerer authentication key as BASE64 encoded string
-     * @param agent                 Agent details
-     * @param script                Script that need to be executed
-     * @return                      Shell execution output
-     * @throws IOException          Request post exception
-     * @throws JSONException        JSON parsing exception
-     */
-    private Content sendShellCommand(String tinkererHost, String authenticationKey, JSONObject agent, String script)
-            throws IOException, JSONException {
-        return Request.Post(
-                tinkererHost + "/test-plan/" + agent.getString("testPlanId") +
-                        "/agent/" + agent.getString("instanceName") + "/operation"
-        ).addHeader(HttpHeaders.AUTHORIZATION, authenticationKey)
-                .addHeader(HttpHeaders.CONTENT_TYPE, "application/json")
-                .bodyString("{\"request\":\"" + script + "\",\"code\":\"SHELL\"}", ContentType.APPLICATION_JSON).
-                        execute().returnContent();
     }
 
     /**
